requirements.txt
pandas
streamlit
altair
matplotlib
seaborn
google-generativeai
